Keep only one shortcut recorder open at once
